20/08 10:00 - 21/08 01:00
 Zo goed als heel het “treasure” gedeelte erin geïmplenteerd samen met de statistieken van users, voedselbanken en algemeen website statistieken. Hier dan ook meteen cron jobs gelinkt die worden gerunned door laravel scheduler. Leaderboards systeem gemaakt(wekelijkse en all time) + pagina. Hier dan ook meteen een scheduler voor gemaakt. Facebook share functie erin gestoken + geprobeerd dat dan meteen de hoeveelheid van iets wordt geshared aka “ik heb x hoeveelheid voedsel gedonneert”. Niet gelukt en lijkt mij niet echt mogelijk, morgen verder daarin kijken. Verder uitgewerkt dat als een user een goodiebag afleverd zonder een account te hebben dat deze hierna een account kan aanmaken om deze “treasures” te collecten. Show pagina van foodbank uitgebreid naar de mogelijkheid om de info aan te passen als de eigenaar hiervan is ingelogt.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -27,7 +27,7 @@ class CreateUsersTable extends Migration
             $table->boolean('isFoodbank')->nullable();
             $table->timestamp('email_verified_at')
                     ->nullable();
-            $table->float('treasures')->nullable();
+            $table->decimal('treasures',10,3)->nullable();
             $table->string('password');
             $table->decimal('lat', 10,7)->nullable();
             $table->decimal('lng',10,7)->nullable();
